adjust the spawn probability better

+++ src/main/java/com/gmail/guitaekm/enderlakes/ConfigInstance.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record ConfigInstance (
         int nrLakes,
-        int powerDistance,
+        double powerDistance,
         List<Integer> cycleWeights,
         int minimumDistance
 ) {
